+++ src/components/Checkbox/index.test.tsx
@@ -1,0 +1,14 @@
+import { render, screen } from "@testing-library/react";
+
+import "@testing-library/jest-dom";
+import { Checkbox } from ".";
+
+describe("ui/Checkbox", () => {
+  it("title is exist", () => {
+    render(<Checkbox />);
+
+    const title = screen.getByText(/this is ui of Checkbox/);
+
+    expect(title).toBeInTheDocument();
+  });
+});
