+++ src/components/Checkbox/index.ts
@@ -1,0 +1,7 @@
+import { Checkbox } from "./index.tsx";
+
+import type { CheckboxProps } from "./index.tsx";
+
+export { Checkbox };
+
+export type { CheckboxProps };
