--- src/components/Sample/index.css.ts
@@ -1,5 +1,0 @@
-import { style } from '@vanilla-extract/css'
-
-export const root = style({
-  color: 'red',
-})
