--- src/components/Sample/index.tsx
@@ -1,15 +1,0 @@
-import 'modern-normalize/modern-normalize.css'
-import '../../reset.css'
-import type { FC } from 'react'
-
-import { clsx } from 'clsx'
-
-import { root } from './index.css'
-
-export type SampleProps = {
-  className?: string
-}
-
-export const Sample: FC<SampleProps> = ({ className }) => (
-  <p className={clsx(root, className)}>this is ui of Sample </p>
-)
